+++ MultifariousProgram/MultifariousProgramModel.java
@@ -2,50 +2,54 @@ package campus_console_project_leader_side.hankyungsoo.MultifariousProgram;
 
 public class MultifariousProgramModel {
 
-    private double result;
-    private int floor;
+	private double result;
+	private int floor;
 
-    //구구단 연산 메소드
-    public int[] run(int number) {
-        int[] result = new int[9];
-        for (int i = 0; i < 9; i++) {
-            result[i] = number * (i + 1);
-        }
-        return result;
-    }
+	// 구구단 연산 메소드
+	public int[] run(int number) {
+		int[] result = new int[9];
+		for (int i = 0; i < 9; i++) {
+			result[i] = number * (i + 1);
+		}
+		return result;
+	}
 
-    // 계산기 연산 메소드
-    public void add(double num1, double num2) {
-        result = num1 + num2;
-    }
+	// 계산기 연산 메소드
+	public void add(double num1, double num2) {
+		result = num1 + num2;
+	}
 
-    public void subtract(double num1, double num2) {
-        result = num1 - num2;
-    }
+	public void subtract(double num1, double num2) {
+		result = num1 - num2;
+	}
 
-    public void multiply(double num1, double num2) {
-        result = num1 * num2;
-    }
+	public void multiply(double num1, double num2) {
+		result = num1 * num2;
+	}
 
-    public void divide(double num1, double num2) {
-        if (num2 != 0) {
-            result = num1 / num2;
-        } else {
-            result = 0;
-        }
-    }
+	public void divide(double num1, double num2) {
+		if (num2 != 0) {
+			result = num1 / num2;
+		} else {
+			result = 0;
+		}
+	}
 
-    public double getResult() {
-        return result;
-    }
+	public double getResult() {
+		return result;
+	}
 
-    // 피라미드 층수 리턴
-    public void setFloor(int floor) {
-        this.floor = floor;
-    }
+	public void setResult() {
+		result = 0;
+	}
 
-    public int getFloor() {
-        return floor;
-    }
+	// 피라미드 층수 리턴
+	public void setFloor(int floor) {
+		this.floor = floor;
+	}
+
+	public int getFloor() {
+		return floor;
+	}
 
 }
